// src/clarity-angular/wigfrid/components/FlexGrid/Grid/enum/RowColFlags.d.ts
/**
 * Flags that specify the state of a grid row or column.
 */
export declare enum RowColFlags {
    /** The row or column is visible. */
    Visible = 1,
    /** The row or column can be resized. */
    AllowResizing = 2,
    /** The row or column can be dragged to a new position with the mouse. */
    AllowDragging = 4,
    /** The row or column can contain merged cells. */
    AllowMerging = 8,
    /** The column can be sorted by clicking its header with the mouse. */
    AllowSorting = 16,
    /** The column was generated automatically. */
    AutoGenerated = 32,
    /** The group row is collapsed. */
    Collapsed = 64,
    /** The row has a parent group that is collapsed. */
    ParentCollapsed = 128,
    /** The row or column is selected. */
    Selected = 256,
    /** The row or column is read-only (cannot be edited). */
    ReadOnly = 512,
    /** Cells in this row or column contain HTML text. */
    HtmlContent = 1024,
    /** Cells in this row or column may contain wrapped text. */
    WordWrap = 2048,
    /** Default settings for new rows. */
    RowDefault = 3,
    /** Default settings for new columns. */
    ColumnDefault = 23,
}